requirements.txt
requests>=2.31.0
pyyaml>=6.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
cryptography>=41.0.0

